turtling: Take Classes from "pq.py':  Class BytesTerminal, ChordsKeyboard, TurtleClientWas
The Git Diffs lose synch after a few Screens of Lines and come out messy
But I made a point of landing most of my Changes in Tops of Files, before they lose synch

I did teach Cued_Py_Lines_Text to drop the paragraphs that we've entirely commented out

I left behind a forwarding address over to 'bin/turtling.py' from:  bin/pq turtle

I didn't leave forwarding addresses behind at:  pq em, pq vi, pq em vi
